+++ raven/__init__.py
@@ -1,5 +1,1 @@
-# -*- coding: utf-8 -*-
-
-"""Top-level package for Raven."""
-
 from .__version__ import __author__, __email__, __version__  # noqa: F401