+++ raven/processes/wps_raven_gr4j_cemaneige.py
@@ -51,4 +51,4 @@ class RavenGR4JCemaNeigeProcess(RavenProcess):
 
     inputs = [wio.ts, wio.nc_spec, params, wio.start_date, wio.end_date, wio.nc_index, wio.duration, wio.run_name,
               wio.name, wio.area, wio.latitude, wio.longitude, wio.elevation, wio.evaporation,
-              wio.rain_snow_fraction]
+              wio.rain_snow_fraction, wio.rvc]